requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
pydantic-settings
sqlalchemy
pymysql
passlib[bcrypt]
python-jose[cryptography]
chromadb
google-generativeai
pypdf
passlib[bcrypt]
bcrypt==4.0.1